Layered the classes for calendar package, please check if this is valid and can perhaps move in the api package created.
