+++ src/main/java/use_case/WeeklyPlanner.java
@@ -1,4 +1,9 @@
-package CalenderPackage;
+package src.main.java.use_case;
+
+import src.main.java.entity.PlanItem;
+import src.main.java.entity.Assignment;
+import src.main.java.entity.Course;
+import src.main.java.entity.Event;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
